Working test, robot staying in place
I got it so the right heel of the robot stays in the exact same place over the floor. Note that the original creator for some reason has a model or animation scaling applied and it was messing things up for. I now have that value defined in code to match.