Walking code is LIVE!
It walks almost perfectly, tracking with the surface as it walks! There is a little bit of glitching going on, but I think that is during the handoff of points that there is a little *backward* movement before the point successfully transitions to the next. Just a consequence of not lifting the foot enough during animation. I can probably fix it