Add proper walking speed based on animation walking (#2)
* Creating develop version

* Playing around with some functions

Nothing fundamentally changed

* Started adding feature, added function for bone calculation

* Added debug printing so I can find relevant values

* Working on adding velocity offset based on foot movement

* Working test, robot staying in place

I got it so the right heel of the robot stays in the exact same place over the floor. Note that the original creator for some reason has a model or animation scaling applied and it was messing things up for. I now have that value defined in code to match.

* Started work on finding the lowest tracker point

I need to switch points to whichever is actually touching the ground. The velocity will then be based off of this point. Note that I need a handoff of sorts so the velocity doesn't jump around when the position changes during swap.

* Added function to find lowest tracking point

I have removed the velocity adjust for now. Once I have verified it correctly finds the point touching the ground, I will re-implement velocity changes.

* Debugged latest function

There were a few errors related to the assignment, but that is fixed. The game is reporting the correct tracker point at each location

* Some cleanup of code, handoff of tracker points

The velocity calculation is still messed up though, the point handoff doesn't work the way I expected and it still has forward and back velocity.

* Walking code is LIVE!

It walks almost perfectly, tracking with the surface as it walks! There is a little bit of glitching going on, but I think that is during the handoff of points that there is a little *backward* movement before the point successfully transitions to the next. Just a consequence of not lifting the foot enough during animation. I can probably fix it

* Debugging running

Running has the feet lift entirely off the ground, so I have it set to a pre-determined speed when it is flying. For some reason though this isn't working.

* Trying to figure out movement state signals and How to recieve

To fix the run animation I have to get the current movement state, and only apply my custom movement velocity if it is currently walking

* Ready for pull request

Walking is working, and my code does not conflict with any of the other animations.

* Added player skeleton and removed buggy grapple code

It appears like the grapple is a work in progress, so this was removed to get the game working. The developer is free to add that back when it works, but for this merged it is removed. Game works, with walk cycle addition.